WishListPageを分割
